feat: move subscription form into shared SubscriptionForm and add shared SubscribeMinimalistic
