+++ src/components/shared/subscribe-minimalistic/index.js
@@ -1,0 +1,3 @@
+import SubscribeMinimalistic from './subscribe-minimalistic';
+
+export default SubscribeMinimalistic;
