+++ src/components/shared/subscription-form/index.js
@@ -1,0 +1,3 @@
+import SubscriptionForm from './subscription-form';
+
+export default SubscriptionForm;
