+++ src/pages/jobs.jsx
@@ -3,13 +3,13 @@ import React from 'react';
 import Hero from 'components/pages/jobs/hero';
 import JobsList from 'components/pages/jobs/jobs-list';
 import Layout from 'components/shared/layout';
-import Subscribe from 'components/shared/subscribe';
+import SubscribeMinimalistic from 'components/shared/subscribe-minimalistic';
 
 const HomePage = () => (
   <Layout headerTheme="black">
     <Hero />
     <JobsList />
-    <Subscribe />
+    <SubscribeMinimalistic />
   </Layout>
 );
